requirements.txt
Django==5.0.13
psycopg2-binary==2.9.9
djangorestframework==3.15.2
pytest==8.3.2
pytest-django==4.8.0
drf-spectacular==0.28.0
PyJWT==2.8.0
flake8==7.1.0
black==23.3.0
python-dotenv==1.0.0
djangorestframework-simplejwt==5.5.0
eventlet==0.39.1
whitenoise==6.6.0
gunicorn==21.2.0
celery==5.3.6
redis==5.0.1
django-redis==5.4.0
Pillow==10.2.0
django-filter==24.1